[BOLT][RISCV] Implement basic instrumentation
This patch implements support for instrumenting direct function calls
and jumps on RISC-V (anything that relies on `createInstrIncMemory`) as
well as the necessary runtime library support (init, fini, and syscall
wrappers).

Only `createInstrIncMemory` is implemented by this patch. However, most
of the other `create*` functions are called when instrumenting  even the
simplest binary on RISC-V and their default implementation triggers an
assert. Therefore, dummy implementations are added by this patch.

Atomic instructions (used to increment counters) are only available on
RISC-V when the A extension is used. A check was added to err when
trying to instrument a binary that was compiled without this extension
enabled.

The syscall wrappers in the runtime library have been implemented by
leveraging the fact that the syscall calling convention matches the user
space one (at least when up to six scalars are passed as arguments): for
every wrapper, a forward declaration is defined in C whereas its
implementation is defined in assembly and simply loads the syscall
number and uses `ecall`. This way, the compiler will prepare all the
arguments in the correct registers when calling the function, saving us
from a lot of repetitive work. The implementation is done in global
assembly (instead of inline assembly in a C function) to make sure the
compiler cannot optimize away the call and skip preparing the argument
registers.

I've taken a slightly different approach than the existing targets with
respect to the data types used by syscalls. I'm not quite sure why
`common.h` defines some system types but I believe it is fine to just
use the system headers for this (we only want to avoid linking against
system libraries; using their headers should not cause problems). For
this reason, I've made sure the definitions in `common.h` are not used
for RISC-V.